requirements.txt
django>=4.2,<5.0
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
pytz>=2024.1
whitenoise>=6.6
gunicorn>=21.2
python-dotenv>=1.0
anthropic>=0.30
